requirements.txt
flask
openai
psycopg2-binary